feat: add UpdatePartnerStatus Azure function
Implement UpdatePartnerStatus timer triggered Azure function. The function is set to run every 6 hours. It identifies kit statuses that have not been updated in the partner status system, calls the partner API to update the status of each kit, verifies the update success with the partner status system, and marks the local kit status event as updated.